new projects/requirements.txt
# Core dependencies
streamlit==1.32.0
python-dotenv==1.0.0

# Audio processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1
noisereduce==3.0.0

# Speech-to-Text
openai-whisper==20231117
torch==2.2.0
torchaudio==2.2.0

# Speaker Diarization
pyannote.audio==3.1.1
speechbrain==0.5.16

# Emotion Detection
transformers==4.38.0
datasets==2.17.0

# Video processing
moviepy==1.0.3
ffmpeg-python==0.2.0

# Utilities
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.2
scipy==1.11.4

# File handling
python-multipart==0.0.9

# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Batch Processing
tqdm==4.66.2
scikit-learn==1.3.2
